modlib/dlfcn:unify same code
