+++ binfmt/libelf/libelf_load.c
@@ -270,15 +270,6 @@ static inline int elf_loadfile(FAR struct elf_loadinfo_s *loadinfo)
     {
       FAR Elf_Shdr *shdr = &loadinfo->shdr[i];
 
-      /* SHF_ALLOC indicates that the section requires memory during
-       * execution.
-       */
-
-      if ((shdr->sh_flags & SHF_ALLOC) == 0)
-        {
-          continue;
-        }
-
 #ifdef CONFIG_ARCH_USE_SEPARATED_SECTION
       if (loadinfo->ehdr.e_type == ET_REL)
         {
