+++ libs/libc/dlfcn/lib_dlclose.c
@@ -25,210 +25,12 @@
 #include <nuttx/config.h>
 
 #include <dlfcn.h>
-#include <assert.h>
-#include <debug.h>
-#include <errno.h>
 
-#include <nuttx/module.h>
 #include <nuttx/lib/modlib.h>
-
-#include "libc.h"
 
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
-
-/****************************************************************************
- * Name: dlremove
- *
- * Description:
- *   Remove a previously installed shared library from memory.
- *
- * Input Parameters:
- *   handle - The shared library handle previously returned by dlopen().
- *
- * Returned Value:
- *   Zero (OK) on success.  On any failure, -1 (ERROR) is returned the
- *   errno value is set appropriately.
- *
- ****************************************************************************/
-
-#ifdef CONFIG_BUILD_PROTECTED
-static inline int dlremove(FAR void *handle)
-{
-  FAR struct module_s *modp = (FAR struct module_s *)handle;
-  void (**array)(void);
-  int ret;
-  int i;
-
-  DEBUGASSERT(modp != NULL);
-
-  /* Get exclusive access to the module registry */
-
-  modlib_registry_lock();
-
-  /* Verify that the module is in the registry */
-
-  ret = modlib_registry_verify(modp);
-  if (ret < 0)
-    {
-      serr("ERROR: Failed to verify module: %d\n", ret);
-      goto errout_with_lock;
-    }
-
-#if CONFIG_MODLIB_MAXDEPEND > 0
-  /* Refuse to remove any module that other modules may depend upon. */
-
-  if (modp->dependents > 0)
-    {
-      serr("ERROR: Module has dependents: %d\n", modp->dependents);
-      ret = -EBUSY;
-      goto errout_with_lock;
-    }
-#endif
-
-  /* Is there an uninitializer? */
-
-  if (modp->modinfo.uninitializer != NULL)
-    {
-      /* Try to uninitialize the module */
-
-      ret = modp->modinfo.uninitializer(modp->modinfo.arg);
-
-      /* Did the module successfully uninitialize? */
-
-      if (ret < 0)
-        {
-          serr("ERROR: Failed to uninitialize the module: %d\n", ret);
-          goto errout_with_lock;
-        }
-
-      /* Nullify so that the uninitializer cannot be called again */
-
-      modp->modinfo.uninitializer = NULL;
-
-      /* Call any .fini_array entries in reverse order */
-
-      array = (void (**)(void))modp->finiarr;
-      for (i = (modp->nfini - 1); i >= 0; i--)
-        {
-          array[i]();
-        }
-
-#if defined(CONFIG_FS_PROCFS) && !defined(CONFIG_FS_PROCFS_EXCLUDE_MODULE)
-      modp->initializer           = NULL;
-      modp->modinfo.arg           = NULL;
-      modp->modinfo.exports       = NULL;
-      modp->modinfo.nexports      = 0;
-#endif
-    }
-
-  /* Release resources held by the module */
-
-  /* Dynamic shared objects have text and data allocated in one
-   * operation to keep the relative positions between the two
-   * areas relative otherwise references to the GOT will fail
-   */
-
-  if (!modp->dynamic)
-    {
-#ifdef CONFIG_ARCH_USE_SEPARATED_SECTION
-      int i;
-
-      for (i = 0; i < modp->nsect && modp->sectalloc[i] != NULL; i++)
-        {
-#  ifdef CONFIG_ARCH_USE_TEXT_HEAP
-          if (up_textheap_heapmember(modp->sectalloc[i]))
-            {
-              up_textheap_free(modp->sectalloc[i]);
-              continue;
-            }
-#  endif
-
-#  ifdef CONFIG_ARCH_USE_DATA_HEAP
-          if (up_dataheap_heapmember(modp->sectalloc[i]))
-            {
-              up_dataheap_free(modp->sectalloc[i]);
-              continue;
-            }
-#  endif
-
-          lib_free(modp->sectalloc[i]);
-        }
-
-      lib_free(modp->sectalloc);
-      modp->sectalloc = NULL;
-      modp->nsect = 0;
-#else
-      if (modp->textalloc != NULL)
-        {
-          /* Free the module memory */
-
-#if defined(CONFIG_ARCH_USE_TEXT_HEAP)
-          up_textheap_free((FAR void *)modp->textalloc);
-#else
-          lib_free((FAR void *)modp->textalloc);
-#endif
-        }
-
-      if (modp->dataalloc != NULL)
-        {
-          /* Free the module memory */
-
-#if defined(CONFIG_ARCH_USE_DATA_HEAP)
-          up_dataheap_free((FAR void *)modp->dataalloc);
-#else
-          lib_free((FAR void *)modp->dataalloc);
-#endif
-        }
-#endif
-    }
-  else
-    {
-      lib_free((FAR void *)modp->textalloc);
-    }
-
-  /* Nullify so that the memory cannot be freed again */
-
-  modp->textalloc = NULL;
-  modp->dataalloc = NULL;
-#if defined(CONFIG_FS_PROCFS) && !defined(CONFIG_FS_PROCFS_EXCLUDE_MODULE)
-  modp->textsize  = 0;
-  modp->datasize  = 0;
-#endif
-
-  /* Free the modules exported symmbols table */
-
-  modlib_freesymtab(modp);
-
-  /* Remove the module from the registry */
-
-  ret = modlib_registry_del(modp);
-  if (ret < 0)
-    {
-      serr("ERROR: Failed to remove the module from the registry: %d\n",
-           ret);
-      goto errout_with_lock;
-    }
-
-#if CONFIG_MODLIB_MAXDEPEND > 0
-  /* Eliminate any dependencies that this module has on other modules */
-
-  modlib_undepend(modp);
-#endif
-  modlib_registry_unlock();
-
-  /* And free the registry entry */
-
-  lib_free(modp);
-  return OK;
-
-errout_with_lock:
-  modlib_registry_unlock();
-  set_errno(-ret);
-  return ERROR;
-}
-#endif
 
 /****************************************************************************
  * Public Functions
@@ -277,15 +79,11 @@ errout_with_lock:
 
 int dlclose(FAR void *handle)
 {
-#if defined(CONFIG_BUILD_FLAT)
+#if defined(CONFIG_BUILD_FLAT) || defined(CONFIG_BUILD_PROTECTED)
   /* In the FLAT build, a shared library is essentially the same as a kernel
    * module.
-   */
-
-  return rmmod(handle);
-
-#elif defined(CONFIG_BUILD_PROTECTED)
-  /* The PROTECTED build is equivalent to the FLAT build EXCEPT that there
+   *
+   * The PROTECTED build is equivalent to the FLAT build EXCEPT that there
    * must be two copies of the module logic:  One residing in kernel
    * space and using the kernel symbol table and one residing in user space
    * using the user space symbol table.
@@ -293,7 +91,7 @@ int dlclose(FAR void *handle)
    * dlremove() is essentially a clone of rmmod().
    */
 
-  return dlremove(handle);
+  return modlib_remove(handle);
 
 #else /* if defined(CONFIG_BUILD_KERNEL) */
   /* The KERNEL build is considerably more complex:  In order to be shared,
