+++ libs/libc/dlfcn/lib_dlopen.c
@@ -24,128 +24,16 @@
 
 #include <nuttx/config.h>
 
-#include <stdlib.h>
-#include <string.h>
 #include <libgen.h>
 #include <dlfcn.h>
-#include <assert.h>
-#include <debug.h>
-#include <errno.h>
-#include <unistd.h>
-#include <sys/param.h>
 
 #include <nuttx/envpath.h>
-#include <nuttx/module.h>
 #include <nuttx/lib/modlib.h>
-
-#include "libc.h"
-#include "debug.h"
+#include <nuttx/lib/lib.h>
 
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
-
-/****************************************************************************
- * Name: dldump_loadinfo
- ****************************************************************************/
-
-#ifdef CONFIG_BUILD_PROTECTED
-#  if defined(CONFIG_DEBUG_INFO) && defined(CONFIG_DEBUG_BINFMT)
-static void dldump_loadinfo(FAR struct mod_loadinfo_s *loadinfo)
-{
-  int i;
-
-  binfo("LOAD_INFO:\n");
-  binfo("  textalloc:    %08lx\n", (long)loadinfo->textalloc);
-  binfo("  datastart:    %08lx\n", (long)loadinfo->datastart);
-  binfo("  textsize:     %ld\n",   (long)loadinfo->textsize);
-  binfo("  datasize:     %ld\n",   (long)loadinfo->datasize);
-  binfo("  filelen:      %ld\n",   (long)loadinfo->filelen);
-  binfo("  filfd:        %d\n",    loadinfo->filfd);
-  binfo("  symtabidx:    %d\n",    loadinfo->symtabidx);
-  binfo("  strtabidx:    %d\n",    loadinfo->strtabidx);
-
-  binfo("ELF Header:\n");
-  binfo("  e_ident:      %02x %02x %02x %02x\n",
-    loadinfo->ehdr.e_ident[0], loadinfo->ehdr.e_ident[1],
-    loadinfo->ehdr.e_ident[2], loadinfo->ehdr.e_ident[3]);
-  binfo("  e_type:       %04x\n",  loadinfo->ehdr.e_type);
-  binfo("  e_machine:    %04x\n",  loadinfo->ehdr.e_machine);
-  binfo("  e_version:    %08x\n",  loadinfo->ehdr.e_version);
-  binfo("  e_entry:      %08lx\n", (long)loadinfo->ehdr.e_entry);
-  binfo("  e_phoff:      %d\n",    loadinfo->ehdr.e_phoff);
-  binfo("  e_shoff:      %d\n",    loadinfo->ehdr.e_shoff);
-  binfo("  e_flags:      %08x\n" , loadinfo->ehdr.e_flags);
-  binfo("  e_ehsize:     %d\n",    loadinfo->ehdr.e_ehsize);
-  binfo("  e_phentsize:  %d\n",    loadinfo->ehdr.e_phentsize);
-  binfo("  e_phnum:      %d\n",    loadinfo->ehdr.e_phnum);
-  binfo("  e_shentsize:  %d\n",    loadinfo->ehdr.e_shentsize);
-  binfo("  e_shnum:      %d\n",    loadinfo->ehdr.e_shnum);
-  binfo("  e_shstrndx:   %d\n",    loadinfo->ehdr.e_shstrndx);
-
-  if (loadinfo->shdr && loadinfo->ehdr.e_shnum > 0)
-    {
-      for (i = 0; i < loadinfo->ehdr.e_shnum; i++)
-        {
-          FAR Elf_Shdr *shdr = &loadinfo->shdr[i];
-          binfo("Sections %d:\n", i);
-#    ifdef CONFIG_ARCH_USE_SEPARATED_SECTION
-          if (loadinfo->ehdr.e_type == ET_REL)
-            {
-              binfo("  sh_alloc:     %08jx\n",
-                    (uintmax_t)loadinfo->sectalloc[i]);
-            }
-#    endif
-
-          binfo("  sh_name:      %08x\n", shdr->sh_name);
-          binfo("  sh_type:      %08x\n", shdr->sh_type);
-          binfo("  sh_flags:     %08x\n", shdr->sh_flags);
-          binfo("  sh_addr:      %08x\n", shdr->sh_addr);
-          binfo("  sh_offset:    %d\n",   shdr->sh_offset);
-          binfo("  sh_size:      %d\n",   shdr->sh_size);
-          binfo("  sh_link:      %d\n",   shdr->sh_link);
-          binfo("  sh_info:      %d\n",   shdr->sh_info);
-          binfo("  sh_addralign: %d\n",   shdr->sh_addralign);
-          binfo("  sh_entsize:   %d\n",   shdr->sh_entsize);
-        }
-    }
-
-  if (loadinfo->phdr && loadinfo->ehdr.e_phnum > 0)
-    {
-      for (i = 0; i < loadinfo->ehdr.e_phnum; i++)
-        {
-          FAR Elf32_Phdr *phdr = &loadinfo->phdr[i];
-          binfo("Program Header %d:\n", i);
-          binfo("  p_type:       %08x\n", phdr->p_type);
-          binfo("  p_offset:     %08x\n", phdr->p_offset);
-          binfo("  p_vaddr:      %08x\n", phdr->p_vaddr);
-          binfo("  p_paddr:      %08x\n", phdr->p_paddr);
-          binfo("  p_filesz:     %08x\n", phdr->p_filesz);
-          binfo("  p_memsz:      %08x\n", phdr->p_memsz);
-          binfo("  p_flags:      %08x\n", phdr->p_flags);
-          binfo("  p_align:      %08x\n", phdr->p_align);
-        }
-    }
-}
-#  else
-#    define dldump_loadinfo(i)
-#  endif
-#endif
-
-/****************************************************************************
- * Name: dldump_initializer
- ****************************************************************************/
-
-#if defined(CONFIG_BUILD_PROTECTED) && defined(CONFIG_MODLIB_DUMPBUFFER)
-static void dldump_initializer(mod_initializer_t initializer,
-                               FAR struct mod_loadinfo_s *loadinfo)
-{
-  modlib_dumpbuffer("Initializer code", (FAR const uint8_t *)initializer,
-                    MIN(loadinfo->textsize - loadinfo->ehdr.e_entry, 512));
-}
-#else
-#  define dldump_initializer(b,l)
-#endif
 
 /****************************************************************************
  * Public Functions
@@ -173,148 +61,7 @@ static void dldump_initializer(mod_initializer_t initializer,
  *
  ****************************************************************************/
 
-#ifdef CONFIG_BUILD_PROTECTED
-/* The PROTECTED build is equivalent to the FLAT build EXCEPT that there
- * must be two copies of the module logic:  One residing in kernel
- * space and using the kernel symbol table and one residing in user space
- * using the user space symbol table.
- *
- * dlinsert() is essentially a clone of insmod().
- */
-
-static inline FAR void *dlinsert(FAR const char *filename)
-{
-  struct mod_loadinfo_s loadinfo;
-  FAR struct module_s *modp;
-  mod_initializer_t initializer;
-  void (**array)(void);
-  int ret;
-  int i;
-
-  binfo("Loading file: %s\n", filename);
-
-  /* Get exclusive access to the module registry */
-
-  modlib_registry_lock();
-
-  /* Initialize the ELF library to load the program binary. */
-
-  ret = modlib_initialize(filename, &loadinfo);
-  dldump_loadinfo(&loadinfo);
-  if (ret != 0)
-    {
-      serr("ERROR: Failed to initialize to load module: %d\n", ret);
-      goto errout_with_loadinfo;
-    }
-
-  /* Allocate a module registry entry to hold the module data */
-
-  modp = (FAR struct module_s *)lib_zalloc(sizeof(struct module_s));
-  if (modp == NULL)
-    {
-      ret = -ENOMEM;
-      binfo("Failed to initialize for load of ELF program: %d\n", ret);
-      goto errout_with_loadinfo;
-    }
-
-  /* Load the program binary */
-
-  ret = modlib_load(&loadinfo);
-  dldump_loadinfo(&loadinfo);
-  if (ret != 0)
-    {
-      binfo("Failed to load ELF program binary: %d\n", ret);
-      goto errout_with_registry_entry;
-    }
-
-  /* Bind the program to the kernel symbol table */
-
-  ret = modlib_bind(modp, &loadinfo);
-  if (ret != 0)
-    {
-      binfo("Failed to bind symbols program binary: %d\n", ret);
-      goto errout_with_load;
-    }
-
-  /* Save the load information */
-
-#ifdef CONFIG_ARCH_USE_SEPARATED_SECTION
-  modp->sectalloc = (FAR void *)loadinfo.sectalloc;
-  modp->nsect = loadinfo.ehdr.e_shnum;
-#endif
-
-  modp->textalloc = (FAR void *)loadinfo.textalloc;
-  modp->dataalloc = (FAR void *)loadinfo.datastart;
-  modp->dynamic   = (loadinfo.ehdr.e_type == ET_DYN);
-#if defined(CONFIG_FS_PROCFS) && !defined(CONFIG_FS_PROCFS_EXCLUDE_MODULE)
-  modp->textsize  = loadinfo.textsize;
-  modp->datasize  = loadinfo.datasize;
-#endif
-
-  /* Get the module initializer entry point */
-
-  initializer = (mod_initializer_t)(loadinfo.textalloc +
-                                    loadinfo.ehdr.e_entry);
-#if defined(CONFIG_FS_PROCFS) && !defined(CONFIG_FS_PROCFS_EXCLUDE_MODULE)
-  modp->initializer = initializer;
-#endif
-  dldump_initializer(initializer, &loadinfo);
-
-  /* Call the module initializer */
-
-  switch (loadinfo.ehdr.e_type)
-    {
-      case ET_REL :
-          ret = initializer(&modp->modinfo);
-          if (ret < 0)
-            {
-              binfo("Failed to initialize the module: %d\n", ret);
-              goto errout_with_load;
-            }
-          break;
-      case ET_DYN :
-
-          /* Process any preinit_array entries */
-
-          array = (void (**)(void)) loadinfo.preiarr;
-          for (i = 0; i < loadinfo.nprei; i++)
-            {
-              array[i]();
-            }
-
-          /* Process any init_array entries */
-
-          array = (void (**)(void)) loadinfo.initarr;
-          for (i = 0; i < loadinfo.ninit; i++)
-            {
-              array[i]();
-            }
-
-          modp->finiarr = loadinfo.finiarr;
-          modp->nfini = loadinfo.nfini;
-          break;
-    }
-
-  /* Add the new module entry to the registry */
-
-  modlib_registry_add(modp);
-
-  modlib_uninitialize(&loadinfo);
-  modlib_registry_unlock();
-  return modp;
-
-errout_with_load:
-  modlib_unload(&loadinfo);
-  modlib_undepend(modp);
-errout_with_registry_entry:
-  lib_free(modp);
-errout_with_loadinfo:
-  modlib_uninitialize(&loadinfo);
-  modlib_registry_unlock();
-  set_errno(-ret);
-  return NULL;
-}
-#elif defined(CONFIG_BUILD_FLAT)
+#if defined(CONFIG_BUILD_FLAT) || defined(CONFIG_BUILD_PROTECTED)
 /* In the FLAT build, a shared library is essentially the same as a kernel
  * module.
  *
@@ -322,6 +69,13 @@ errout_with_loadinfo:
  * - No automatic binding of symbols
  * - No dependencies
  * - mode is ignored.
+ *
+ * The PROTECTED build is equivalent to the FLAT build EXCEPT that there
+ * must be two copies of the module logic:  One residing in kernel
+ * space and using the kernel symbol table and one residing in user space
+ * using the user space symbol table.
+ *
+ * dlinsert() is essentially a clone of insmod().
  */
 
 static inline FAR void *dlinsert(FAR const char *filename)
@@ -341,7 +95,7 @@ static inline FAR void *dlinsert(FAR const char *filename)
    * name.
    */
 
-  handle = insmod(filename, basename(name));
+  handle = modlib_insert(filename, basename(name));
   lib_free(name);
   return handle;
 }
