+++ libs/libc/dlfcn/lib_dlsym.c
@@ -25,86 +25,12 @@
 #include <nuttx/config.h>
 
 #include <dlfcn.h>
-#include <assert.h>
-#include <debug.h>
-#include <errno.h>
 
-#include <nuttx/symtab.h>
-#include <nuttx/module.h>
 #include <nuttx/lib/modlib.h>
 
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
-
-/****************************************************************************
- * Name: dlgetsym
- *
- * Description:
- *   dlgetsym() implements dlsym() for the PROTECTED build.
- *
- * Input Parameters:
- *   handle - The opaque, non-NULL value returned by a previous successful
- *            call to insmod().
- *   name   - A pointer to the symbol name string.
- *
- * Returned Value:
- *   See dlsym().
- *
- ****************************************************************************/
-
-#ifdef CONFIG_BUILD_PROTECTED
-static inline FAR const void *dlgetsym(FAR void *handle,
-                                       FAR const char *name)
-{
-  FAR struct module_s *modp = (FAR struct module_s *)handle;
-  FAR const struct symtab_s *symbol;
-  int err;
-  int ret;
-
-  /* Verify that the module is in the registry */
-
-  modlib_registry_lock();
-  ret = modlib_registry_verify(modp);
-  if (ret < 0)
-    {
-      serr("ERROR: Failed to verify module: %d\n", ret);
-      err = -ret;
-      goto errout_with_lock;
-    }
-
-  /* Does the module have a symbol table? */
-
-  if (modp->modinfo.exports == NULL || modp->modinfo.nexports == 0)
-    {
-      serr("ERROR: Module has no symbol table\n");
-      err = ENOENT;
-      goto errout_with_lock;
-    }
-
-  /* Search the symbol table for the matching symbol */
-
-  symbol = symtab_findbyname(modp->modinfo.exports, name,
-                             modp->modinfo.nexports);
-  if (symbol == NULL)
-    {
-      serr("ERROR: Failed to find symbol in symbol \"%s\" in table\n", name);
-      err = ENOENT;
-      goto errout_with_lock;
-    }
-
-  /* Return the address within the module associated with the symbol */
-
-  modlib_registry_unlock();
-  DEBUGASSERT(symbol->sym_value != NULL);
-  return symbol->sym_value;
-
-errout_with_lock:
-  modlib_registry_unlock();
-  set_errno(err);
-  return NULL;
-}
-#endif
 
 /****************************************************************************
  * Public Functions
@@ -143,15 +69,11 @@ errout_with_lock:
 
 FAR void *dlsym(FAR void *handle, FAR const char *name)
 {
-#if defined(CONFIG_BUILD_FLAT)
+#if defined(CONFIG_BUILD_FLAT) || defined(CONFIG_BUILD_PROTECTED)
   /* In the FLAT build, a shared library is essentially the same as a kernel
    * module.
-   */
-
-  return (FAR void *)modsym(handle, name);
-
-#elif defined(CONFIG_BUILD_PROTECTED)
-  /* The PROTECTED build is equivalent to the FLAT build EXCEPT that there
+   *
+   * The PROTECTED build is equivalent to the FLAT build EXCEPT that there
    * must be two copies of the module logic:  One residing in kernel
    * space and using the kernel symbol table and one residing in user space
    * using the user space symbol table.
@@ -159,7 +81,7 @@ FAR void *dlsym(FAR void *handle, FAR const char *name)
    * dlgetsem() is essentially a clone of modsym().
    */
 
-  return (FAR void *)dlgetsym(handle, name);
+  return (FAR void *)modlib_getsymbol(handle, name);
 
 #else /* if defined(CONFIG_BUILD_KERNEL) */
   /* The KERNEL build is considerably more complex:  In order to be shared,
