+++ libs/libc/modlib/CMakeLists.txt
@@ -36,7 +36,11 @@ if(CONFIG_LIBC_MODLIB)
     modlib_symtab.c
     modlib_uninit.c
     modlib_unload.c
-    modlib_verify.c)
+    modlib_verify.c
+    modlib_gethandle.c
+    modlib_getsymbol.c
+    modlib_insert.c
+    modlib_remove.c)
 
   list(APPEND SRCS modlib_globals.S)
