+++ libs/libc/modlib/Make.defs
@@ -26,6 +26,8 @@ CSRCS += modlib_bind.c modlib_depend.c modlib_init.c modlib_iobuffer.c
 CSRCS += modlib_load.c modlib_loadhdrs.c modlib_verify.c
 CSRCS += modlib_read.c modlib_registry.c modlib_sections.c
 CSRCS += modlib_symbols.c modlib_symtab.c modlib_uninit.c modlib_unload.c
+CSRCS += modlib_gethandle.c modlib_getsymbol.c modlib_insert.c
+CSRCS += modlib_remove.c
 
 # Add the modlib directory to the build
