+++ libs/libc/modlib/modlib_unload.c
@@ -72,39 +72,40 @@ int modlib_unload(FAR struct mod_loadinfo_s *loadinfo)
           if (up_textheap_heapmember((FAR void *)loadinfo->sectalloc[i]))
             {
               up_textheap_free((FAR void *)loadinfo->sectalloc[i]);
-              continue;
             }
+          else
 #  endif
 
 #  ifdef CONFIG_ARCH_USE_DATA_HEAP
           if (up_dataheap_heapmember((FAR void *)loadinfo->sectalloc[i]))
             {
               up_dataheap_free((FAR void *)loadinfo->sectalloc[i]);
-              continue;
             }
+          else
 #  endif
-
-          lib_free((FAR void *)loadinfo->sectalloc[i]);
+            {
+              lib_free((FAR void *)loadinfo->sectalloc[i]);
+            }
         }
 
       lib_free(loadinfo->sectalloc);
 #else
       if (loadinfo->textalloc != 0)
         {
-#if defined(CONFIG_ARCH_USE_TEXT_HEAP)
+#  if defined(CONFIG_ARCH_USE_TEXT_HEAP)
           up_textheap_free((FAR void *)loadinfo->textalloc);
-#else
+#  else
           lib_free((FAR void *)loadinfo->textalloc);
-#endif
+#  endif
         }
 
       if (loadinfo->datastart != 0)
         {
-#if defined(CONFIG_ARCH_USE_DATA_HEAP)
+#  if defined(CONFIG_ARCH_USE_DATA_HEAP)
           up_dataheap_free((FAR void *)loadinfo->datastart);
-#else
+#  else
           lib_free((FAR void *)loadinfo->datastart);
-#endif
+#  endif
         }
 #endif
     }
