+++ sched/module/mod_rmmod.c
@@ -26,12 +26,6 @@
 
 #include <nuttx/config.h>
 
-#include <sys/types.h>
-#include <assert.h>
-#include <debug.h>
-#include <errno.h>
-
-#include <nuttx/kmalloc.h>
 #include <nuttx/module.h>
 #include <nuttx/lib/modlib.h>
 
@@ -58,154 +52,7 @@
 
 int rmmod(FAR void *handle)
 {
-  FAR struct module_s *modp = (FAR struct module_s *)handle;
-  int ret;
-
-  DEBUGASSERT(modp != NULL);
-
-  /* Get exclusive access to the module registry */
-
-  modlib_registry_lock();
-
-  /* Verify that the module is in the registry */
-
-  ret = modlib_registry_verify(modp);
-  if (ret < 0)
-    {
-      berr("ERROR: Failed to verify module: %d\n", ret);
-      goto errout_with_lock;
-    }
-
-#if CONFIG_MODLIB_MAXDEPEND > 0
-  /* Refuse to remove any module that other modules may depend upon. */
-
-  if (modp->dependents > 0)
-    {
-      berr("ERROR: Module has dependents: %d\n", modp->dependents);
-      ret = -EBUSY;
-      goto errout_with_lock;
-    }
-#endif
-
-  /* Is there an uninitializer? */
-
-  if (modp->modinfo.uninitializer != NULL)
-    {
-      /* Try to uninitialize the module */
-
-      ret = modp->modinfo.uninitializer(modp->modinfo.arg);
-
-      /* Did the module successfully uninitialize? */
-
-      if (ret < 0)
-        {
-          berr("ERROR: Failed to uninitialize the module: %d\n", ret);
-          goto errout_with_lock;
-        }
-
-      /* Nullify so that the uninitializer cannot be called again */
-
-      modp->modinfo.uninitializer = NULL;
-#if defined(CONFIG_FS_PROCFS) && !defined(CONFIG_FS_PROCFS_EXCLUDE_MODULE)
-      modp->initializer           = NULL;
-      modp->modinfo.arg           = NULL;
-      modp->modinfo.exports       = NULL;
-      modp->modinfo.nexports      = 0;
-#endif
-    }
-
-  /* Release resources held by the module */
-
-  if (modp->textalloc != NULL || modp->dataalloc != NULL)
-    {
-      /* Free the module memory and nullify so that the memory cannot
-       * be freed again
-       *
-       * NOTE: For dynamic shared objects there is only a single
-       * allocation: the text/data were allocated in one operation
-       */
-
-      if (!modp->dynamic)
-        {
-#ifdef CONFIG_ARCH_USE_SEPARATED_SECTION
-          int i;
-
-          for (i = 0; i < modp->nsect && modp->sectalloc[i] != NULL; i++)
-            {
-#  ifdef CONFIG_ARCH_USE_TEXT_HEAP
-              if (up_textheap_heapmember(modp->sectalloc[i]))
-                {
-                  up_textheap_free(modp->sectalloc[i]);
-                  continue;
-                }
-#  endif
-
-#  ifdef CONFIG_ARCH_USE_DATA_HEAP
-              if (up_dataheap_heapmember(modp->sectalloc[i]))
-                {
-                  up_dataheap_free(modp->sectalloc[i]);
-                  continue;
-                }
-#  endif
-
-              kmm_free(modp->sectalloc[i]);
-            }
-
-          kmm_free(modp->sectalloc);
-          modp->sectalloc = NULL;
-          modp->nsect = 0;
-#else
-#  if defined(CONFIG_ARCH_USE_TEXT_HEAP)
-          up_textheap_free((FAR void *)modp->textalloc);
-#  else
-          kmm_free((FAR void *)modp->textalloc);
-#  endif
-#  if defined(CONFIG_ARCH_USE_DATA_HEAP)
-          up_dataheap_free((FAR void *)modp->dataalloc);
-#  else
-          kmm_free((FAR void *)modp->dataalloc);
-#  endif
-#endif
-        }
-      else
-        {
-          kmm_free((FAR void *)modp->textalloc);
-        }
-
-      modp->textalloc = NULL;
-      modp->dataalloc = NULL;
-#if defined(CONFIG_FS_PROCFS) && !defined(CONFIG_FS_PROCFS_EXCLUDE_MODULE)
-      modp->textsize  = 0;
-      modp->datasize  = 0;
-#endif
-    }
-
-  /* Remove the module from the registry */
-
-  ret = modlib_registry_del(modp);
-  if (ret < 0)
-    {
-      berr("ERROR: Failed to remove the module from the registry: %d\n",
-           ret);
-      goto errout_with_lock;
-    }
-
-#if CONFIG_MODLIB_MAXDEPEND > 0
-  /* Eliminate any dependencies that this module has on other modules */
-
-  modlib_undepend(modp);
-#endif
-  modlib_registry_unlock();
-
-  /* And free the registry entry */
-
-  kmm_free(modp);
-  return OK;
-
-errout_with_lock:
-  modlib_registry_unlock();
-  set_errno(-ret);
-  return ERROR;
+  return modlib_remove(handle);
 }
 
 #endif /* CONFIG_MODULE */
